renaming file. added notes about possible enhancements
